# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SPMS API",
        "description": "subscription-platform-management-system API Documentation",
        "version": "API v1.0"
    },
    "paths": {
...
        "/api/v1/packages": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get packages",
                "description": "Get Packages",
                "operationId": "PackagesIndex",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page.",
                        "required": false,
                        "schema": {
                            "type": "numeric"
                        },
                        "example": "5"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "numeric"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "HBO"
                    },
                    {
                        "name": "sku",
                        "in": "query",
                        "description": "sku",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "SKU1234"
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "active",
                        "required": false,
                        "schema": {
                            "type": "bool"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PackageResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/packages/{id}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get Package",
                "description": "Get package by ID",
                "operationId": "GetPackage",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PackageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Edit Package",
                "operationId": "PackageUpdate",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PackageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Delete Package",
                "operationId": "PackageDelete",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package ID",
                        "required": true,
                        "schema": {
                            "type": "numeric"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/packages/": {
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Create Package",
                "operationId": "PackageStore",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PackageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PackageRequest": {
                "required": [
                    "name",
                    "sku",
                    "image_url",
                    "active"
                ],
                "properties": {
                    "name": {
                        "description": "package name",
                        "format": "string",
                        "example": "HBO",
                        "nullable": false
                    },
                    "image_url": {
                        "description": "image_url",
                        "format": "string",
                        "example": "https://via.placeholder.com/640x480.png/00aa22?text=illum",
                        "nullable": false
                    },
                    "sku": {
                        "description": "sku",
                        "format": "string",
                        "example": "SKU1234",
                        "nullable": false
                    },
                    "active": {
                        "description": "active",
                        "format": "boolean",
                        "example": "1",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "PackageResource": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64",
                        "default": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Package name",
                        "format": "string",
                        "default": "Package 1"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku",
                        "format": "string",
                        "default": "PACKAGE1234"
                    },
                    "image_url": {
                        "title": "image_url",
                        "description": "Package image url",
                        "format": "string",
                        "default": "https://via.placeholder.com/640x480.png/0011dd?text=eos"
                    },
                    "active": {
                        "title": "active",
                        "description": "Package active",
                        "format": "boolean",
                        "default": true
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Package created at",
                        "format": "string",
                        "default": "2024-01-01 00:00:00"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Package updated at",
                        "format": "string",
                        "default": "2024-01-01 00:00:00"
                    },
                    "deleted_at": {
                        "title": "deleted_at",
                        "description": "Package deleted at",
                        "format": "string",
                        "default": "2024-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    }
}